Create new API object and refactor full export
This is so much better: You don't need to create a blob for each file in advance; you can do it with a single call to create a new tree. This also removes the need for all this background process handoff nonsense. Now we can do it all pretty quickly in only a handful of API calls and harmonize the cron and cli methods. Lastly, exporting twice with no changes doesn't add an extra commit.