Rewrite the valid check as a while loop.
Using recursion for the valid loop was hitting the max stack depth with XDebug on. I don't know if this is a common problem, but I suspect this loop will hit a max stack depth in enough cases that we should avoid it.

The while loop could be written a bit more elegantly, but this does the job with minimal alterations.