Rewrite caching object to use Transients.
The other alternative would be to write a better method for distributing this data across many/several/lots of database rows, but that seems like a lot of database hits when we can allow the object cache to manage this for us. WordPress falls back to using the database for transients if there is no persistent object cache installed, so letting WordPress handle all this for us makes sense.

Fixes #72.