Add dependency on container to Api object
This required a cascading set of changes, many of which we'll attempt to back out later, but a lot of things currently depend on the Api being instantiable, which it isn't if you don't have access to the container. Unfortunately, the Post object generally doesn't, so we have to pass in the Api object everywhere.