Clean up global state and properly isolate tests
We were getting the log messages because the
instance was hooking in on boot, which also deletes
all the posts currently in the database, firing the
`delete_post` hook, which runs our real code,
because the real code is hooked into real hooks
that are really running.

These changes remove the need for the global
variable, splits off the hooking step into a separate
function and hook that we can then remove before
running our tests. This way, none of the real code
gets hooked in and runs during testing.