Fully unit test response object
We're unit testing it the best we can, but we can't test
wp_send_json_* functions because they kill the script. Also
simplify the API of the response object, so everything passes
solely through `success` and `error` which defers to the logging
and send a JSON response if in the correct context.