Add wpghs_pre_import_meta filter
Provides point for plugin developers to hook in and modify
the post meta before it gets imported. This filter can be used
to save some of the data elsewhere for the provided Post,
and should then return only the remaining meta that should be
saved to post_meta specifically.

Fixes #106