Add "best guess" path deletion
On the off-chance the paths returned from GitHub don't
match the path saved for the post in the database, we can
try a "best guess" to extract the title from the pathname.
While we can't "know" that the path matches the title,
these are the defaults, so we can try to file the title.

Readds code lost in 765e493e518da9af1bdf715d17a5cc2dcc489e95

(See GitHub comment for original code.)