Fix test to create and use post
If a WPGHS_Post was passed an ID that didn't exist in
the database, it would still be treated as new even if
it shouldn't have been. Once this bug was fixed, it
exposed this problem in the tests, which assumed we
could pretend that if a Post is passed an ID, it exists
in the DB. This resolves that inconsistency.