Allow customizations when importing posts
The three actions that this adds allow us to do initializations
before all the posts are saved, then post-specific modifications
for each post before it gets saved, and finally we can restore
settings after all the posts are saved.

This was needed in order to be able to track the current language
used in WPML, then set the active language for each post (because
otherwise post categories would get deleted), and finally reset
the active lanugage to the one remembered in the first action.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>