Only update posts that are marked as modified
Github tells us which files were modified. If we find that information in the
webhook callback, let's not blindly update all posts but instead only update
those posts that we were told were modified.

This adds a new function to the payload class to get the actual head commit
included in the payload.

json_decode(json_encode(), true) is used to convert the object to an array -
I'm not sure if there is a different way that you would like to use for this
(I'm not really a PHP developer - but this works).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>